backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
pydantic>=2.7
pytest>=8.0
httpx>=0.27
rdkit-pypi>=2023.3.1b1
